+++ modules/group-test.mjs
@@ -1,3 +1,6 @@
+import GMToolkit from "./gm-toolkit.mjs"
+import { getGroup } from "./utility.mjs"
+
 export async function launchGroupTest (groupOptions, testParameters) {
   new game.gmtoolkit.grouptest({ groupOptions, testParameters }).render(true)
 }
@@ -32,7 +35,6 @@ export async function getGroupMembers (groupType = game.settings.get("wfrp4e-gm-
     playerGroup: game.gmtoolkit.utility.getGroup(groupType).map(g => g.uuid),
     selected: game.gmtoolkit.utility.getGroup("company", { interaction: "selected", present: true }).map(g => g.uuid),
     npcTokens: game.gmtoolkit.utility.getGroup("npcTokens").map(g => g.document.uuid),
-    // _game.gmtoolkit.utility.getGroup("tokens", { interaction: "selected" })
     controlled: canvas.tokens.placeables.filter(t => t.controlled & t.actor.type !== "vehicle").map(g => g.document.uuid)
   }
   return members
@@ -42,15 +44,50 @@ export async function getGroupMembers (groupType = game.settings.get("wfrp4e-gm-
 export async function runGroupTest (testSkill, testOptions) {
   await game.settings.set("wfrp4e-gm-toolkit", "aggregateResultGroupTest", [])
   let actorTestResult = ""
+  const activePlayers = getGroup("assigned", { active: true })
+  const targetGroup = (testOptions.targetGroup.filter(member => member !== null))
 
-  for (const member of testOptions.targetGroup) {
-    if (member === null) continue
+  for (const member of targetGroup) {
     // Make sure to get the actor rather than the token document
     let actor = await fromUuid(member)
     actor = actor?.actor ? actor.actor : actor
-    actorTestResult = await runActorTest(actor, testSkill, testOptions)
+
+    if (testOptions?.difficulty === undefined) {
+      // Overrides default difficulty to Average depending on module setting and combat state
+      let difficulty = "challenging"
+      if (game.settings.get("wfrp4e", "testDefaultDifficulty") && (game.combat != null))
+        difficulty = game.combat.started ? "challenging" : "average"
+      else if (game.settings.get("wfrp4e", "testDefaultDifficulty"))
+        difficulty = "average"
+      testOptions.difficulty = difficulty
+    }
+
+    // Delegate tests to active assigned players
+    if (!testOptions.bypass
+      && activePlayers.map(p => p.character.uuid).includes(member)) {
+
+      const requestData = {
+        type: "requestRoll",
+        payload: {
+          character: actor,
+          test: {
+            type: "skill",
+            against: testSkill,
+            options: testOptions
+          }
+        }
+      }
+      await game.socket.emit(`module.${GMToolkit.MODULE_ID}`, requestData)
+    } else {
+      // Don't delegate if character player is not active or roll dialog is not bypassed
+      actorTestResult = await runActorTest(actor, testSkill, testOptions)
+    }
   }
-  return sendAggregateGroupTestResults(testSkill, testOptions)
+
+  if (game.settings.get("wfrp4e-gm-toolkit", "aggregateResultGroupTest").length > 0) {
+    sendAggregateGroupTestResults(testSkill, testOptions)
+  }
+
 }
 
 
@@ -78,7 +115,8 @@ async function sendAggregateGroupTestResults (testSkill, testOptions) {
     actorTestResultMessage
       += `<strong>${testResult.actor.name}:</strong> `
       + `<abbr title="${testResult.description}"><strong>${testResult.sl} SL</strong></abbr> (${testResult.roll} v ${testResult.target})`
-    if (testResult.characteristic) actorTestResultMessage += ` [${game.i18n.localize(testResult.characteristic.abrev)}]`
+    if (testResult.characteristic) actorTestResultMessage += ` [${game.i18n.localize(game.wfrp4e.config.characteristics[testResult.characteristic])
+    }]`
 
     // Transfer to group test message and nuke actor test message
     groupTestResultsMessage += `${actorTestResultMessage}</br>`
@@ -89,16 +127,22 @@ async function sendAggregateGroupTestResults (testSkill, testOptions) {
     content: groupTestResultsMessage,
     whisper: game.users.filter(u => u.isGM).map(u => u.id)
   })
+
+  // clean up aggregate results pseudo-setting
+  await game.settings.set("wfrp4e-gm-toolkit", "aggregateResultGroupTest", [])
 }
 
 
-async function runActorTest (actor, testSkill, testOptions) {
+export async function runActorTest (actor, testSkill, testOptions) {
   let actorSkill = game.gmtoolkit.utility.hasSkill(actor, testSkill, "silent")
   let setupData = {
     bypass: testOptions.bypass,
-    testModifier: testOptions.testModifier,
-    rollMode: testOptions.rollMode,
-    absolute: { difficulty: testOptions.difficulty },
+    fields: {
+      modifier: testOptions?.testModifier || 0,
+      rollMode: testOptions.rollMode,
+      // TODO: set difficulty selection to system default when bypassing dialog
+      difficulty: testOptions.difficulty
+    },
     groupTest: true,
     title: game.i18n.format("GMTOOLKIT.Dialog.MakeSecretGroupTest.RollTitle", { skill: actorSkill?.name, actor: actor.name })
   }
@@ -121,6 +165,7 @@ async function runActorTest (actor, testSkill, testOptions) {
         = (actorSkill.advanced.value === "adv")
           ? game.settings.get("wfrp4e-gm-toolkit", "fallbackAdjustDifficulty")
           : 0
+          // TODO: Refactor to use wfrp4e.utility.alterDifficulty
       const difficultySetting
         = (actorSkill.advanced.value === "adv" && stepAdjustDifficulty !== 0)
           ? { modify: { difficulty: stepAdjustDifficulty } }
@@ -143,24 +188,57 @@ async function runActorTest (actor, testSkill, testOptions) {
   }
 }
 
+/**
+ * Incrementally updates a world-scoped setting after each actor test
+ * @param {Object} actorTestResult
+ * @returns {Object} setting: aggregateResultGroupTest
+ */
+export async function updateGroupTestResults (actorTestResult) {
+  await game.settings.set(
+    "wfrp4e-gm-toolkit",
+    "aggregateResultGroupTest",
+    actorTestResult
+  )
+
+  return GMToolkit.log(
+    true,
+    "updateGroupTestResults",
+    game.settings.get(GMToolkit.MODULE_ID, "aggregateResultGroupTest")
+  )
+}
 
 // Intercept test result
 Hooks.on("wfrp4e:rollTest", async function (testData, chatData) {
   if (testData.options.groupTest) {
     const groupTestResult = []
-    groupTestResult.push(...game.settings.get("wfrp4e-gm-toolkit", "aggregateResultGroupTest"))
+    await groupTestResult.push(...game.settings.get("wfrp4e-gm-toolkit", "aggregateResultGroupTest"))
 
     const testResult = {
       actor: testData.token || testData.actor,
       skill: testData?.skill,
-      characteristic: testData?.characteristic,
+      // characteristic: ( testData?.characteristic ? testData?.characteristicKey ),
       outcome: testData.outcome,
       sl: testData.result.SL,
       description: testData.result.description,
       roll: testData.result.roll,
       target: testData.target
     }
+
+    if (testData?.characteristic) (
+      testResult.characteristic = testData.characteristicKey
+    )
+
     await groupTestResult.push(testResult)
-    return await game.settings.set("wfrp4e-gm-toolkit", "aggregateResultGroupTest", groupTestResult)
+    // GMToolkit.log(true, "groupTestResult: this test result", testResult)
+    // GMToolkit.log(true, "groupTestResult: groupTestResult", groupTestResult)
+
+    if (game.user.isUniqueGM) {
+      return await updateGroupTestResults(groupTestResult)
+    } else {
+      return game.socket.emit(`module.${GMToolkit.MODULE_ID}`,
+        { type: "aggregateGroupTestResults", payload: groupTestResult })
+    }
+
   }
 })
+